stable without WITH_SERIAL_PROBE
also done with memcpubusy rdcpu_finished as clocks, they are sampled
on negedge psram_clk now.

this branch is branched from before hdmi clock introduction as that was
adding its own instabilities which are very hard to deal with all at the
same time.

timing closure still isn't looking great, but works well.

this is all on the 5" ST7262 display. the scaler skips 1 out of 6 hsync
pulses. it's not compatible with the 7" that allegedly uses HX8264.